+++ cmd/graph_generator/main.go
@@ -28,8 +28,8 @@ var (
 	methodsNotExported = types.SetWith(
 		"ArgMinMax", "Broadcast", "BroadcastInDim",
 		"BatchNormForInference", "BatchNormForTraining", "BatchNormGradient",
-		"Concatenate", "ConvGeneralDilated", "DotGeneral", "FFT", "Gather", "Iota",
-		"ReduceMax", "ReduceMin", "ReduceProduct", "ReduceSum", "ReduceWindow",
+		"Concatenate", "ConvertDType", "ConvGeneralDilated", "DotGeneral", "FFT", "Gather", "Iota",
+		"ReduceAnd", "ReduceMax", "ReduceMin", "ReduceOr", "ReduceProduct", "ReduceSum", "ReduceWindow",
 		"Reshape", "Reverse", "RngBitGenerator",
 		"ScatterAdd", "SelectAndScatterSum", "SelectAndScatterMax", "SelectAndScatterMin",
 		"Sign", "Slice",
@@ -79,10 +79,21 @@ func buildMethodInfo() (methods []*MethodInfo) {
 				case "...Op":
 					pi.BackendType = "...backends.Op"
 					pi.GraphType = "...*Node"
-					mi.OpInputsList = paramName
+					mi.OpInputSlices = append(mi.OpInputSlices, paramName)
 					pi.NodeInputType = "[]*Node"
 					pi.CopyStatement = fmt.Sprintf("slices.Clone(%s)", paramName)
 					pi.ConvertStatement = fmt.Sprintf("xslices.Map(%s, func(node *Node) backends.Op { return node.outputOps[0] })...", paramName)
+					pi.Format = "[#%s]"
+					pi.FormatValue = fmt.Sprintf(
+						`strings.Join(xslices.Map(ni.%s, func (node *Node) string { return fmt.Sprintf("#%%d", node.Id()) }), ", ")`,
+						paramName)
+				case "[]Op":
+					pi.BackendType = "[]backends.Op"
+					pi.GraphType = "[]*Node"
+					mi.OpInputSlices = append(mi.OpInputSlices, paramName)
+					pi.NodeInputType = "[]*Node"
+					pi.CopyStatement = fmt.Sprintf("slices.Clone(%s)", paramName)
+					pi.ConvertStatement = fmt.Sprintf("xslices.Map(%s, func(node *Node) backends.Op { return node.outputOps[0] })", paramName)
 					pi.Format = "[#%s]"
 					pi.FormatValue = fmt.Sprintf(
 						`strings.Join(xslices.Map(ni.%s, func (node *Node) string { return fmt.Sprintf("#%%d", node.Id()) }), ", ")`,
@@ -131,7 +142,7 @@ func buildMethodInfo() (methods []*MethodInfo) {
 					pi.FormatValue = "ni." + pi.Name
 				}
 			}
-			mi.HasGraph = mi.OpInputsList == "" && len(mi.OpInputs) == 0
+			mi.HasGraph = len(mi.OpInputSlices) == 0 && len(mi.OpInputs) == 0
 
 		}
 		for _, field := range funcInfo.Type.Results.List {
@@ -151,7 +162,7 @@ type MethodInfo struct {
 	BackendName, GraphName string
 	HasGraph               bool
 	OpInputs               []string
-	OpInputsList           string
+	OpInputSlices          []string
 	Inputs                 []*ParameterInfo
 	Exported, Excluded     bool
 	Comments               []string
@@ -227,14 +238,13 @@ Outputs: */}}{{if not .HasMultipleOutputs}}node *Node{{/*
 
 Body: */}}{
 {{if .HasGraph}}	g.AssertBuilding()
-{{else}}{{if ne .OpInputsList ""}}	g := validateBuildingGraphFromInputs({{.OpInputsList}}...)
-{{else}}	g := validateBuildingGraphFromInputs({{range .OpInputs}}{{.}}, {{end}})
-{{end}}{{end}}	inputs := &nodeInputs{{.BackendName}}{
+{{else}}	inputNodes := []*Node{ {{range .OpInputs}}{{.}}, {{end}} }
+{{range .OpInputSlices}}	inputNodes = append(inputNodes, {{.}}...)
+{{end}}	g := validateBuildingGraphFromInputs(inputNodes...)
+{{end}}	inputs := &nodeInputs{{.BackendName}}{
 {{range .Inputs}}		{{.Name}}: {{.CopyStatement}},		
 {{end}}	}
-{{if not .HasGraph}}	{{if eq .OpInputsList ""}}inputNodes := []*Node{ {{range .OpInputs}}{{.}}, {{end}} } 
-{{else}}	inputNodes := {{.OpInputsList}}
-{{end}}{{end}}{{/*
+{{/*
 
 Convert result(s) to node(s):